% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.e.R
\name{kernel.e}
\alias{kernel.e}
\title{Epanechnikov kernel}
\usage{
kernel.e(u)
}
\arguments{
\item{u}{vector of observations}
}
\value{
evaluation of the Epanechnikov kernel
}
\description{
Evaluates the Epanechnikov kernel
}
\examples{
x = rnorm(100,1,2)
kernel.e(x)
}
