% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.g.R
\name{kernel.g}
\alias{kernel.g}
\title{Gaussian kernel}
\usage{
kernel.g(u)
}
\arguments{
\item{u}{vector of observations}
}
\value{
evaluation of the Gaussian kernel
}
\description{
Evaluates the Gaussian kernel
}
\examples{
x = rnorm(100,1,2)
kernel.g(x)
}
