% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likbox.R
\name{likbox}
\alias{likbox}
\title{Likelihood function of the BoxCox transformation}
\usage{
likbox(h, data, n)
}
\arguments{
\item{h}{parameter of the Box-Cox transformation}

\item{data}{joint vector of controls (first) and cases}

\item{n}{length of the vector of controls}
}
\value{
the likelihood function of the BoxCox transformation
}
\description{
Computation of the likelihood function of the BoxCox transformation
}
\examples{
h=-1.6
controls=rnorm(50,6,1)
cases=rnorm(100,6.5,0.5)
likbox(h,c(controls,cases),n=length(controls))
}
