% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OVL.D.R
\name{OVL.D}
\alias{OVL.D}
\title{OVL.D}
\usage{
OVL.D(x, y, alpha = 0.05)
}
\arguments{
\item{x}{controls}

\item{y}{cases}

\item{alpha}{confidence level}
}
\value{
confidence interval
}
\description{
Parametric approach using the delta method
}
\examples{
controls = rnorm(50,6,1)
cases = rnorm(100,6.5,0.5)
OVL.D (controls,cases)
}
