% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OVL.LogitBCAN.R
\name{OVL.LogitBCAN}
\alias{OVL.LogitBCAN}
\title{OVL.LogitBCAN}
\usage{
OVL.LogitBCAN(x, y, alpha = 0.05, B = 100, h_ini = -0.6)
}
\arguments{
\item{x}{controls}

\item{y}{cases}

\item{alpha}{confidence level}

\item{B}{bootstrap size}

\item{h_ini}{initial value in the optimization problem}
}
\value{
confidence interval
}
\description{
BCAN procedure carried out in the logit scale and back-transformed
}
\examples{
controls = rnorm(50,6,1)
cases = rnorm(100,6.5,0.5)
OVL.LogitBCAN (controls,cases)
}
