% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OVL.LogitD.R
\name{OVL.LogitD}
\alias{OVL.LogitD}
\title{OVL.LogitD}
\usage{
OVL.LogitD(x, y, alpha = 0.05)
}
\arguments{
\item{x}{controls}

\item{y}{cases}

\item{alpha}{confidence level}
}
\value{
confidence interval
}
\description{
Parametric approach using the delta method after switching to a logit scale and then transforming back
}
\examples{
controls = rnorm(50,6,1)
cases = rnorm(100,6.5,0.5)
OVL.LogitD (controls,cases)
}
