% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OVL.LogitDBC.R
\name{OVL.LogitDBC}
\alias{OVL.LogitDBC}
\title{OVL.LogitDBC}
\usage{
OVL.LogitDBC(x, y, alpha = 0.05, h_ini = -0.6)
}
\arguments{
\item{x}{controls}

\item{y}{cases}

\item{alpha}{confidence level}

\item{h_ini}{initial value in the optimization problem}
}
\value{
confidence interval
}
\description{
Parametric approach using the delta method after the Box-Cox transformation after switching to a logit scale and then transforming back
}
\examples{
controls = rnorm(50,6,1)
cases = rnorm(100,6.5,0.5)
OVL.LogitDBC (controls,cases)
}
