% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdd.R
\name{ssdd}
\alias{ssdd}
\title{Sample variance computation}
\usage{
ssdd(x)
}
\arguments{
\item{x}{vector of observations}
}
\value{
the sample variance
}
\description{
Computes the sample variance of a vector of observations
}
\examples{
x = rnorm(100,1,2)
ssdd(x)
}
