% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Simulated data with normal distributions to showcase the CI'S Overlap Coefficient (OVL) calculation}
\format{
A data frame with 100 rows and 2 variables:
\describe{
 \item{controls}{Simulated data from a N(10,1)distribution for the control group}
 \item{cases}{Simulated data from a N(10.5,0.5)distribution for the case group}
}
}
\usage{
data(test_data)
}
\description{
Contains controls and cases data from normal distributions
}
\examples{

data(test_data)
}
\references{
This data set was artificially created for the OVL.CI package.
}
\keyword{datasets}
