% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U.R
\name{U}
\alias{U}
\title{Auxiliary function}
\usage{
U(mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{mu1}{sample mean of a vector x}

\item{mu2}{sample mean of a vector y}

\item{sigma1}{sample standard deviation of a vector x}

\item{sigma2}{sample standard deviation of a vector y}
}
\value{
evaluation of an auxiliary function
}
\description{
Evaluates an auxiliary function
}
\examples{
U(1,2,1,1)
}
